Created IDatabase and default implementation
This was done to allow for easy dependency injection
