+++ src/Chirp.Core/DBFacade.cs
@@ -4,57 +4,18 @@ namespace Chirp.Core;
 
 public class DBFacade
 {
-    private readonly string _path = Environment.GetEnvironmentVariable("CHIRPDBPATH") ??
-                                    Path.GetTempPath() + "chirp.db";
+    public IDatabase db { get; }
 
-    private string ConnectionString => $"Data Source={_path}";
-
-    public DBFacade()
+    public DBFacade(IDatabase database)
     {
-        EnsureCreated();
+        db = database;
     }
 
-    public void EnsureCreated()
-    {
-        // check if the database exists, if not call reset
-        if (!File.Exists(_path))
-        {
-            Reset();
-        }
-    }
-
-    private void Reset()
-    {
-        using var connection = new SqliteConnection(ConnectionString);
-        connection.Open();
-
-        using var command = connection.CreateCommand();
-        // call the SQL command to create the table
-
-        command.CommandText = """
-                              drop table if exists user;
-                              create table user (
-                                user_id integer primary key autoincrement,
-                                username string not null,
-                                email string not null,
-                                pw_hash string not null
-                              );
-
-                              drop table if exists message;
-                              create table message (
-                                message_id integer primary key autoincrement,
-                                author_id integer not null,
-                                text string not null,
-                                pub_date integer
-                              );
-                              """;
-
-        command.ExecuteNonQuery();
-    }
+    
 
     public IEnumerable<Cheep> GetCheeps(string? authorUsername = null, int? limit = null)
     {
-        using var connection = new SqliteConnection(ConnectionString);
+        using var connection = new SqliteConnection(db.ConnectionString);
         connection.Open();
 
         using var command = connection.CreateCommand();
