+++ src/Chirp.Core/IDatabase.cs
@@ -2,5 +2,9 @@ namespace Chirp.Core;
 
 public interface IDatabase
 {
+    public string ConnectionString { get; }
     
+    public void EnsureCreated();
+    
+    public void Reset();
 }
